refactor: move pagination logic into usePagination hook
